removed the reference to timeSpent from the SheetView, removed the timeSpent var from the StreakView, introduced the viewModel, moved the half of the logic to the viewModel
